% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{portaldata}
\alias{portaldata}
\alias{portaldata-package}
\title{Portal Data Set and Analysis}
\description{
This package contains data sets from the Portal Project
for easy reuse and reanalysis.
}
