% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{Taxonomy of species sampled}
\format{
A data frame with 10 columns and 54 rows. Columns include: "label", "species_id", "genus", "species", "taxa", "ott_name", "approximate_match", "ott_id",
        "is_synonym", and "flags"
}
\source{
\url{https://lunasare.github.io/spring2023-data-science/data/portal-species-taxonomy.csv}
}
\usage{
taxonomy
}
\description{
A data set containing the taxonomy of species sampled for the Portal Project Data.
We followed the taxonomy from the Open Tree of Life .
}
\keyword{datasets}
